fix: update dashboard status at Reconcile level
