+++ controllers/dashboard_controller.go
@@ -19,6 +19,7 @@ package controllers
 import (
 	"bytes"
 	"context"
+	"crypto/sha256"
 	"encoding/json"
 	"fmt"
 	"net/http"
@@ -160,11 +161,11 @@ func (r *GrafanaDashboardReconciler) Reconcile(ctx context.Context, req ctrl.Req
 		return syncResult, err
 	}
 
-	dashboard := &v1beta1.GrafanaDashboard{}
+	cr := &v1beta1.GrafanaDashboard{}
 	err := r.Client.Get(ctx, client.ObjectKey{
 		Namespace: req.Namespace,
 		Name:      req.Name,
-	}, dashboard)
+	}, cr)
 	if err != nil {
 		if errors.IsNotFound(err) {
 			err = r.onDashboardDeleted(ctx, req.Namespace, req.Name)
@@ -177,22 +178,30 @@ func (r *GrafanaDashboardReconciler) Reconcile(ctx context.Context, req ctrl.Req
 		return ctrl.Result{RequeueAfter: RequeueDelay}, err
 	}
 
-	instances, err := r.GetMatchingDashboardInstances(ctx, dashboard, r.Client)
+	instances, err := r.GetMatchingDashboardInstances(ctx, cr, r.Client)
 	if err != nil {
-		controllerLog.Error(err, "could not find matching instances", "name", dashboard.Name, "namespace", dashboard.Namespace)
+		controllerLog.Error(err, "could not find matching instances", "name", cr.Name, "namespace", cr.Namespace)
 		return ctrl.Result{RequeueAfter: RequeueDelay}, err
 	}
 
 	controllerLog.Info("found matching Grafana instances for dashboard", "count", len(instances.Items))
 
-	dashboardJson, err := r.fetchDashboardJson(dashboard)
+	dashboardJson, err := r.fetchDashboardJson(cr)
 	if err != nil {
-		controllerLog.Error(err, "error fetching dashboard", "dashboard", dashboard.Name)
+		controllerLog.Error(err, "error fetching dashboard", "dashboard", cr.Name)
 		return ctrl.Result{RequeueAfter: RequeueDelay}, nil
 	}
 
+	dashboardFromJson, hash, err := r.prepareDashboardModel(cr, dashboardJson)
+	if err != nil {
+		controllerLog.Error(err, "failed to prepare dashboard model", "dashboard", cr.Name)
+		return ctrl.Result{RequeueAfter: RequeueDelay}, nil
+	}
+
+	uid := fmt.Sprintf("%s", dashboardFromJson["uid"])
+
 	// Garbage collection for a case where dashboard uid get changed, dashboard creation is expected to happen in a separate reconcilication cycle
-	if dashboard.IsUpdatedUID(dashboardJson) {
+	if cr.IsUpdatedUID(uid) {
 		controllerLog.Info("dashboard uid got updated, deleting dashboards with the old uid")
 		err = r.onDashboardDeleted(ctx, req.Namespace, req.Name)
 		if err != nil {
@@ -200,8 +209,8 @@ func (r *GrafanaDashboardReconciler) Reconcile(ctx context.Context, req ctrl.Req
 		}
 
 		// Clean up uid, so further reconcilications can track changes there
-		dashboard.Status.UID = ""
-		err = r.Client.Status().Update(ctx, dashboard)
+		cr.Status.UID = ""
+		err = r.Client.Status().Update(ctx, cr)
 		if err != nil {
 			return ctrl.Result{RequeueAfter: RequeueDelay}, err
 		}
@@ -213,7 +222,7 @@ func (r *GrafanaDashboardReconciler) Reconcile(ctx context.Context, req ctrl.Req
 	success := true
 	for _, grafana := range instances.Items {
 		// check if this is a cross namespace import
-		if grafana.Namespace != dashboard.Namespace && !dashboard.IsAllowCrossNamespaceImport() {
+		if grafana.Namespace != cr.Namespace && !cr.IsAllowCrossNamespaceImport() {
 			continue
 		}
 
@@ -230,24 +239,27 @@ func (r *GrafanaDashboardReconciler) Reconcile(ctx context.Context, req ctrl.Req
 			// first reconcile the plugins
 			// append the requested dashboards to a configmap from where the
 			// grafana reconciler will pick them up
-			err = ReconcilePlugins(ctx, r.Client, r.Scheme, &grafana, dashboard.Spec.Plugins, fmt.Sprintf("%v-dashboard", dashboard.Name))
+			err = ReconcilePlugins(ctx, r.Client, r.Scheme, &grafana, cr.Spec.Plugins, fmt.Sprintf("%v-dashboard", cr.Name))
 			if err != nil {
-				controllerLog.Error(err, "error reconciling plugins", "dashboard", dashboard.Name, "grafana", grafana.Name)
+				controllerLog.Error(err, "error reconciling plugins", "dashboard", cr.Name, "grafana", grafana.Name)
 				success = false
 			}
 		}
 
 		// then import the dashboard into the matching grafana instances
-		err = r.onDashboardCreated(ctx, &grafana, dashboard, dashboardJson)
+		err = r.onDashboardCreated(ctx, &grafana, cr, dashboardFromJson, hash)
 		if err != nil {
-			controllerLog.Error(err, "error reconciling dashboard", "dashboard", dashboard.Name, "grafana", grafana.Name)
+			controllerLog.Error(err, "error reconciling dashboard", "dashboard", cr.Name, "grafana", grafana.Name)
 			success = false
 		}
 	}
 
 	// if the dashboard was successfully synced in all instances, wait for its re-sync period
 	if success {
-		return ctrl.Result{RequeueAfter: dashboard.GetResyncPeriod()}, nil
+		cr.Status.Hash = hash
+		cr.Status.LastResync = metav1.Time{Time: time.Now()}
+		cr.Status.UID = uid
+		return ctrl.Result{RequeueAfter: cr.GetResyncPeriod()}, r.Client.Status().Update(ctx, cr)
 	}
 
 	return ctrl.Result{RequeueAfter: RequeueDelay}, nil
@@ -314,14 +326,9 @@ func (r *GrafanaDashboardReconciler) onDashboardDeleted(ctx context.Context, nam
 	return nil
 }
 
-func (r *GrafanaDashboardReconciler) onDashboardCreated(ctx context.Context, grafana *v1beta1.Grafana, cr *v1beta1.GrafanaDashboard, dashboardJson []byte) error {
+func (r *GrafanaDashboardReconciler) onDashboardCreated(ctx context.Context, grafana *v1beta1.Grafana, cr *v1beta1.GrafanaDashboard, dashboardFromJson map[string]interface{}, hash string) error {
 	if grafana.IsExternal() && cr.Spec.Plugins != nil {
 		return fmt.Errorf("external grafana instances don't support plugins, please remove spec.plugins from your dashboard cr")
-	}
-
-	dashboardJson, err := r.resolveDatasources(cr, dashboardJson)
-	if err != nil {
-		return err
 	}
 
 	grafanaClient, err := client2.NewGrafanaClient(ctx, r.Client, grafana)
@@ -329,25 +336,13 @@ func (r *GrafanaDashboardReconciler) onDashboardCreated(ctx context.Context, gra
 		return err
 	}
 
-	var dashboardFromJson map[string]interface{}
-	err = json.Unmarshal(dashboardJson, &dashboardFromJson)
-	if err != nil {
-		return err
-	}
-
-	uid, _ := dashboardFromJson["uid"].(string) //nolint:errcheck
-	if uid == "" {
-		uid = string(cr.UID)
-	}
-
-	dashboardFromJson["uid"] = uid
-
 	// update/create the dashboard if it doesn't exist in the instance or has been changed
-	hash := cr.Hash(dashboardJson)
+	uid := fmt.Sprintf("%s", dashboardFromJson["uid"])
 	exists, err := r.Exists(grafanaClient, uid)
 	if err != nil {
 		return err
 	}
+
 	if exists && cr.Unchanged(hash) && !cr.ResyncPeriodHasElapsed() {
 		return nil
 	}
@@ -376,18 +371,10 @@ func (r *GrafanaDashboardReconciler) onDashboardCreated(ctx context.Context, gra
 		return errors.NewBadRequest(fmt.Sprintf("error creating dashboard, status was %v", resp.Status))
 	}
 
+	// TODO: do we need to change anything here?
 	// NOTE: When dashboard is uploaded with a new dashboardFromJson["uid"], but with an old name, the old uid is preserved. So, better to rely on the resp.UID here
 	grafana.Status.Dashboards = grafana.Status.Dashboards.Add(cr.Namespace, cr.Name, resp.UID)
-	err = r.Client.Status().Update(ctx, grafana)
-	if err != nil {
-		return err
-	}
-
-	cr.Status.Hash = hash
-	cr.Status.LastResync = metav1.Time{Time: time.Now()}
-	cr.Status.UID = resp.UID
-
-	return r.Client.Status().Update(ctx, cr)
+	return r.Client.Status().Update(ctx, grafana)
 }
 
 // map data sources that are required in the dashboard to data sources that exist in the instance
@@ -435,6 +422,32 @@ func (r *GrafanaDashboardReconciler) fetchDashboardJson(dashboard *v1beta1.Grafa
 	default:
 		return nil, fmt.Errorf("unknown source type %v found in dashboard %v", sourceTypes[0], dashboard.Name)
 	}
+}
+
+// prepareDashboardModel resolves datasources, updates uid (if needed) and converts raw json to type grafana client accepts
+func (r *GrafanaDashboardReconciler) prepareDashboardModel(cr *v1beta1.GrafanaDashboard, dashboardJson []byte) (map[string]interface{}, string, error) {
+	dashboardJson, err := r.resolveDatasources(cr, dashboardJson)
+	if err != nil {
+		return map[string]interface{}{}, "", err
+	}
+
+	hash := sha256.New()
+	hash.Write(dashboardJson)
+
+	var dashboardFromJson map[string]interface{}
+	err = json.Unmarshal(dashboardJson, &dashboardFromJson)
+	if err != nil {
+		return map[string]interface{}{}, "", err
+	}
+
+	uid, _ := dashboardFromJson["uid"].(string) //nolint:errcheck
+	if uid == "" {
+		uid = string(cr.UID)
+	}
+
+	dashboardFromJson["uid"] = uid
+
+	return dashboardFromJson, fmt.Sprintf("%x", hash.Sum(nil)), nil
 }
 
 func (r *GrafanaDashboardReconciler) Exists(client *grapi.Client, uid string) (bool, error) {
